feat(mobile): move new-chat pickers to native formSheet routes
Branch/model/project pickers are now expo-router formSheet routes with
native headers and a headerLeft close button (same for the filter sheet),
mirroring the filter sheet's nested-stack structure. Selection state moved
into newChatPreferencesStore (baseBranch unpersisted via partialize) so
the routes and the widget share it. The @expo/ui picker sheets are
deleted; the context sheet stays in-tree for the attachments context.
