+++ apps/mobile/app/(authenticated)/(home)/new-chat/_layout.tsx
@@ -1,24 +1,18 @@
 import { Stack } from "expo-router";
 import { SheetCloseButton } from "@/screens/(authenticated)/(home)/components/SheetCloseButton";
 
-export default function FilterLayout() {
+export default function NewChatLayout() {
 	return (
 		<Stack
 			screenOptions={{
 				headerBackButtonDisplayMode: "minimal",
 				headerShadowVisible: false,
+				headerLeft: () => <SheetCloseButton />,
 			}}
 		>
-			<Stack.Screen
-				name="index"
-				options={{
-					title: "Filter",
-					headerLeft: () => <SheetCloseButton />,
-				}}
-			/>
+			<Stack.Screen name="branch" options={{ title: "Branch" }} />
+			<Stack.Screen name="model" options={{ title: "Model" }} />
 			<Stack.Screen name="project" options={{ title: "Project" }} />
-			<Stack.Screen name="host" options={{ title: "Host" }} />
-			<Stack.Screen name="sort" options={{ title: "Sort" }} />
 		</Stack>
 	);
 }
