+++ apps/mobile/app/(authenticated)/(home)/new-chat/branch.tsx
@@ -1,0 +1,3 @@
+import { BranchPickerScreen } from "@/screens/(authenticated)/(home)/new-chat/branch";
+
+export default BranchPickerScreen;
