+++ apps/mobile/app/(authenticated)/(home)/new-chat/model.tsx
@@ -1,0 +1,3 @@
+import { ModelPickerScreen } from "@/screens/(authenticated)/(home)/new-chat/model";
+
+export default ModelPickerScreen;
