+++ apps/mobile/app/(authenticated)/(home)/new-chat/project.tsx
@@ -1,0 +1,3 @@
+import { ProjectPickerScreen } from "@/screens/(authenticated)/(home)/new-chat/project";
+
+export default ProjectPickerScreen;
