+++ apps/mobile/screens/(authenticated)/(home)/components/SheetCloseButton/SheetCloseButton.tsx
@@ -1,4 +1,5 @@
 import Ionicons from "@expo/vector-icons/Ionicons";
+import { useRouter } from "expo-router";
 import { Pressable } from "react-native";
 import { useTheme } from "@/hooks/useTheme";
 import { cn } from "@/lib/utils";
@@ -7,10 +8,11 @@ export function SheetCloseButton({
 	onPress,
 	className,
 }: {
-	onPress: () => void;
+	onPress?: () => void;
 	className?: string;
 }) {
 	const theme = useTheme();
+	const router = useRouter();
 	return (
 		<Pressable
 			accessibilityLabel="Close"
@@ -19,7 +21,7 @@ export function SheetCloseButton({
 				className,
 			)}
 			hitSlop={8}
-			onPress={onPress}
+			onPress={onPress ?? (() => router.back())}
 		>
 			<Ionicons name="close" size={20} color={theme.foreground} />
 		</Pressable>
