+++ apps/mobile/screens/(authenticated)/(home)/new-chat/model/ModelPickerScreen.tsx
@@ -1,0 +1,52 @@
+import Ionicons from "@expo/vector-icons/Ionicons";
+import { SUPERSET_CHAT_MODELS } from "@superset/shared/agent-models";
+import { useRouter } from "expo-router";
+import { Pressable, ScrollView } from "react-native";
+import { Text } from "@/components/ui/text";
+import { useTheme } from "@/hooks/useTheme";
+import { useNewChatPreferencesStore } from "@/screens/(authenticated)/(home)/workspaces/components/NewChatWidget/stores/newChatPreferencesStore";
+
+export function ModelPickerScreen() {
+	const router = useRouter();
+	const theme = useTheme();
+	const modelId = useNewChatPreferencesStore((state) => state.modelId);
+	const setModelId = useNewChatPreferencesStore((state) => state.setModelId);
+
+	return (
+		<ScrollView
+			className="bg-background flex-1 px-6"
+			contentContainerStyle={{ flexGrow: 1, paddingVertical: 8 }}
+		>
+			{SUPERSET_CHAT_MODELS.map((model) => {
+				const isSelected = model.id === modelId;
+				return (
+					<Pressable
+						key={model.id}
+						onPress={() => {
+							setModelId(model.id);
+							router.back();
+						}}
+						className="flex-row items-center gap-2.5 py-2.5"
+					>
+						<Text
+							className="flex-1 text-sm font-medium"
+							style={{ color: theme.foreground }}
+						>
+							{model.label}
+						</Text>
+						<Text className="text-xs" style={{ color: theme.mutedForeground }}>
+							{model.provider}
+						</Text>
+						{isSelected ? (
+							<Ionicons
+								name="checkmark-circle"
+								size={18}
+								color={theme.primary}
+							/>
+						) : null}
+					</Pressable>
+				);
+			})}
+		</ScrollView>
+	);
+}
