+++ apps/mobile/screens/(authenticated)/(home)/new-chat/model/index.ts
@@ -1,0 +1,1 @@
+export { ModelPickerScreen } from "./ModelPickerScreen";
