+++ apps/mobile/screens/(authenticated)/(home)/new-chat/project/ProjectPickerScreen.tsx
@@ -1,0 +1,81 @@
+import Ionicons from "@expo/vector-icons/Ionicons";
+import { useRouter } from "expo-router";
+import { Pressable, ScrollView, View } from "react-native";
+import { Text } from "@/components/ui/text";
+import { useHostWorkspaces } from "@/hooks/useHostWorkspaces";
+import { useTheme } from "@/hooks/useTheme";
+import { ProjectAvatar } from "@/screens/(authenticated)/(home)/filter/components/ProjectAvatar";
+import { useNewChatTargets } from "@/screens/(authenticated)/(home)/workspaces/components/NewChatWidget/hooks/useNewChatTargets";
+import { useNewChatPreferencesStore } from "@/screens/(authenticated)/(home)/workspaces/components/NewChatWidget/stores/newChatPreferencesStore";
+
+export function ProjectPickerScreen() {
+	const router = useRouter();
+	const theme = useTheme();
+	const { workspaces } = useHostWorkspaces();
+	const { targets, defaultTarget } = useNewChatTargets(workspaces);
+	const targetKey = useNewChatPreferencesStore((state) => state.targetKey);
+	const setTargetKey = useNewChatPreferencesStore(
+		(state) => state.setTargetKey,
+	);
+
+	const selectedKey =
+		targets.find((target) => target.key === targetKey)?.key ??
+		defaultTarget?.key ??
+		null;
+
+	return (
+		<ScrollView
+			className="bg-background flex-1 px-6"
+			contentContainerStyle={{ flexGrow: 1, paddingVertical: 8 }}
+		>
+			{targets.length === 0 ? (
+				<Text
+					className="py-6 text-center text-sm"
+					style={{ color: theme.mutedForeground }}
+				>
+					No projects on an online host
+				</Text>
+			) : null}
+			{targets.map((target) => {
+				const isSelected = target.key === selectedKey;
+				return (
+					<Pressable
+						key={target.key}
+						onPress={() => {
+							setTargetKey(target.key);
+							router.back();
+						}}
+						className="flex-row items-center gap-2.5 py-2.5"
+					>
+						<ProjectAvatar
+							name={target.projectName}
+							iconUrl={target.projectIconUrl}
+							size={32}
+						/>
+						<View className="flex-1">
+							<Text
+								className="text-sm font-medium"
+								style={{ color: theme.foreground }}
+							>
+								{target.projectName}
+							</Text>
+							<Text
+								className="text-xs"
+								style={{ color: theme.mutedForeground }}
+							>
+								{target.hostName}
+							</Text>
+						</View>
+						{isSelected ? (
+							<Ionicons
+								name="checkmark-circle"
+								size={18}
+								color={theme.primary}
+							/>
+						) : null}
+					</Pressable>
+				);
+			})}
+		</ScrollView>
+	);
+}
