+++ apps/mobile/screens/(authenticated)/(home)/new-chat/project/index.ts
@@ -1,0 +1,1 @@
+export { ProjectPickerScreen } from "./ProjectPickerScreen";
