--- apps/mobile/screens/(authenticated)/(home)/workspaces/components/NewChatWidget/components/BranchPickerSheet/index.ts
@@ -1,1 +1,0 @@
-export { BranchPickerSheet } from "./BranchPickerSheet";
