--- apps/mobile/screens/(authenticated)/(home)/workspaces/components/NewChatWidget/components/ModelPickerSheet/index.ts
@@ -1,1 +1,0 @@
-export { ModelPickerSheet } from "./ModelPickerSheet";
