--- apps/mobile/screens/(authenticated)/(home)/workspaces/components/NewChatWidget/components/TargetPickerSheet/index.ts
@@ -1,1 +1,0 @@
-export { TargetPickerSheet } from "./TargetPickerSheet";
